% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\docType{data}
\name{color_reference}
\alias{color_reference}
\title{Color names, hexadecimal, and CIE Lab colorspace representations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2126 rows and 13 columns.}
\source{
{https://github.com/colorjs/color-namer} and R `colors()`
}
\usage{
color_reference
}
\description{
\describe{
  \item{hex}{hexadecimal color representation (without the # at the beginning)}
  \item{L,a,b}{CIE Lab colorspace representation of `hex`}
  \item{name}{Preferred human-readable name of the color}
  \item{name_alt}{All available human-readable names of the color}
  \item{roygbiv,basic,html,R,pantone,x11,ntc}{Source dataset containing the color}
}
}
\keyword{datasets}
