% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color_reference}
\alias{color_reference}
\title{Color names, hexadecimal, and CIE Lab colorspace representations}
\source{
{https://github.com/colorjs/color-namer} and R \code{colors()}
}
\description{
\describe{
\item{hex}{hexadecimal color representation (without the # at the beginning)}
\item{L,a,b}{CIE Lab colorspace representation of \code{hex}}
\item{name}{Preferred human-readable name of the color}
\item{name_alt}{All available human-readable names of the color}
\item{roygbiv,basic,html,R,pantone,x11,ntc}{Source dataset containing the color}
}
}
\keyword{internal}
