% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pretty_bytes}
\alias{pretty_bytes}
\title{Bytes in a human readable string}
\usage{
pretty_bytes(bytes)
}
\arguments{
\item{bytes}{Numeric vector, number of bytes.}
}
\value{
Character vector, the formatted sizes.
}
\description{
Bytes in a human readable string
}
\examples{
pretty_bytes(1337)
pretty_bytes(133337)
pretty_bytes(13333337)
pretty_bytes(1333333337)
pretty_bytes(133333333337)
}

