% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{pretty_color}
\alias{pretty_color}
\alias{pretty_colour}
\title{Color definition (like RGB) to a name}
\usage{
pretty_color(color)

pretty_colour(color)
}
\arguments{
\item{color}{A scalar color that is usable as an input to \code{col2rgb()}
(assumed to be in the sRGB color space).}
}
\value{
A character string that is the closest named colors to the input
color.  The output will have an attribute of alternate color names (named
"alt").
}
\description{
Color definition (like RGB) to a name
}
